requirements.txt
biopython==1.78
certifi==2020.12.5
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
numba==0.53.0
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.20.4
scipy==1.6.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.59.0
